refactor: xtheme 변수 파일 분리
