+++ packages/user/src/styles/theme/components.ts
@@ -1,0 +1,14 @@
+const container = {
+	center: true,
+	padding: '2rem',
+	screens: {
+		'2xl': '1920px',
+	},
+};
+
+const boxShadow = {
+	md: 'rgba(255,255,255, 0.15) 0px 5px 15px 0px', // toast
+	lg: '0px 0px 20px 0px #ffffffD0', // modal
+};
+
+export { boxShadow, container };
