+++ packages/user/src/styles/theme/index.ts
@@ -1,0 +1,5 @@
+export * from './animations';
+export * from './colors';
+export * from './components';
+export { default as spacing } from './spacing';
+export * from './typography';
