+++ packages/user/src/styles/theme/spacing.ts
@@ -1,0 +1,18 @@
+const spacing = {
+	1: '2px',
+	2: '4px',
+	2.5: '6px',
+	3: '8px',
+	4: '12px',
+	5: '16px',
+	6: '20px',
+	7: '24px',
+	8: '28px',
+	9: '32px',
+	10: '36px',
+	11: '40px',
+	12: '44px',
+	15: '70px',
+};
+
+export default spacing;
